requirements.txt
# 핵심 의존성
pandas>=2.0.0
openpyxl>=3.1.0
requests>=2.31.0
rapidfuzz>=3.6.1
selenium>=4.22.0
webdriver-manager>=4.0.2
python-dotenv>=1.0.1
tenacity>=8.3.0
tqdm>=4.66.4

# 테스트 의존성
pytest>=8.0.0
pytest-cov>=4.1.0
